requirements.txt
flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.2.1
langchain==1.0.2
langchain-community==0.4.1
langchain-openai==1.0.1
langchain-chroma==1.0.0
chromadb>=0.4.24